requirements.txt
aac_metrics==0.5.5
accelerate==1.5.2
audiomentations==0.40.0
datasets==3.5.1
huggingface_hub==0.30.1
importlib_resources==5.12.0
jsonlines==4.0.0
loguru==0.7.3
matplotlib==3.10.1
numpy==1.26.3
einops==0.7.0
timm==0.4.12
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
pandas==2.2.3
peft==0.15.2
PyYAML==6.0.2
Requests==2.32.3
ruamel.base==1.0.0
seaborn==0.13.2
sentence_transformers==4.0.2
SQLAlchemy==2.0.40
sqlmodel==0.0.24
streamlit==1.44.1
torchlibrosa==0.1.0
tqdm==4.67.1
transformers==4.46.3
wandb==0.19.10
yt_dlp==2025.3.31
rouge_score
bert_score
laion-clap
scikit-learn
evaluate
gpustat